api/requirements.txt
flask
python-dotenv
langchain
langchain-community
sqlalchemy
oracledb==1.4.2
ollama
diskcache
llama-index 
llama-index-llms-ollama
llama-index-llms-langchain
fastapi 
uvicorn
sentence-transformers
numpy<2
requests
beautifulsoup4
cx_Oracle

pyhive[hive]
pandas
pyarrow
kerberos
streamlit
watchdog


